Extract utility functions from AdminResourceController to a separate module
