+++ web/ex_admin/controller.ex
@@ -14,4 +14,18 @@ defmodule ExAdmin.Controller do
     Enum.find ExAdmin.get_registered, %{}, &(Map.get(&1, :controller_route) == resource_name)
   end
 
+  defmacro __using__(_opts) do
+    quote do
+      import ExAdmin.Controller
+
+      def set_theme(conn, _) do
+        assign(conn, :theme, ExAdmin.theme)
+      end
+
+      def set_layout(conn, _) do
+        layout = Application.get_env(:ex_admin, :layout) || "#{conn.assigns.theme.name}.html"
+        put_layout(conn, layout)
+      end
+    end
+  end
 end
